+++ vm/src/serde/deserialize_program.rs
@@ -196,11 +196,22 @@ fn arbitrary_parent_location(u: &mut Unstructured, depth: u8) -> arbitrary::Resu
 
 #[cfg_attr(
     all(feature = "arbitrary", feature = "std"),
-    derive(Arbitrary, Clone, Serialize)
+    derive(Arbitrary, Serialize)
 )]
-#[derive(Deserialize, Debug, PartialEq, Eq)]
+#[derive(Deserialize, Debug, PartialEq, Eq, Clone)]
 pub struct DebugInfo {
     instruction_locations: HashMap<usize, InstructionLocation>,
+}
+
+impl DebugInfo {
+    pub fn new(instruction_locations: HashMap<usize, InstructionLocation>) -> Self {
+        Self {
+            instruction_locations,
+        }
+    }
+    pub fn get_instruction_locations(&self) -> HashMap<usize, InstructionLocation> {
+        self.instruction_locations.clone()
+    }
 }
 
 #[cfg_attr(all(feature = "arbitrary", feature = "std"), derive(Arbitrary))]
@@ -214,6 +225,17 @@ pub struct InstructionLocation {
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq)]
 pub struct InputFile {
     pub filename: String,
+}
+
+impl InputFile {
+    #[cfg(feature = "std")]
+    pub fn get_content(&self) -> Result<String, String> {
+        let content = std::fs::read_to_string(self.filename.clone());
+        if let Ok(content) = content {
+            return Ok(content);
+        }
+        Err(format!("Failed to read file {}", self.filename.clone()))
+    }
 }
 
 #[cfg_attr(all(feature = "arbitrary", feature = "std"), derive(Arbitrary))]